Version with DSP filters
I broke my local git.   Hard forcing over to the new version, so I can locally recover.

+++ Feather_TestBox_v0/DisplayHandling.ino
@@ -1,5 +1,5 @@
 
-bool CheckCableStatusByte(byte errorCheck){
+bool CheckCableStatusByte(uint16_t errorCheck){
   return ((cableState.statusByte & errorCheck)==errorCheck);
 }
 /*
